+++ src/i18n/locales/en/badge.json
@@ -1,4 +1,5 @@
 {
+  "all": "All",
   "pending": "Pending",
   "active": "Active",
   "passed": "Passed",
